增加：打卡、取消打卡、订阅、取消订阅、查看订阅与打卡信息接口

+++ backend/src/main/java/com/chinasoft/backend/model/request/EENavigationRequest.java
@@ -11,7 +11,7 @@ public class EENavigationRequest {
 
     private String userLatitude;
     
-    private Integer facilityId;
+    private Long facilityId;
 
     private Integer facilityType;
 }
